feat: product detail main info

+++ app/src/main/java/com/zebrand/app1food30s/ui/offers/ManageOffer.kt
@@ -57,7 +57,7 @@ class ManageOffer : AppCompatActivity() {
                         id,
                         name,
                         imageUrl,
-                        price,
+                        price.toInt(),
                         numProduct.toInt(),
                         date
                     )
